Fixed: People signing out after mensa is closed

+++ app/Http/Controllers/SigninController.php
@@ -295,6 +295,12 @@ class SigninController extends Controller
             return redirect(route('home'));
         }
 
+        // We don't want people to be able to sign out even after mensa is closed. This is annoying for the cooks
+        // when they have already bought stuff with their budget.
+        if($mensa->isClosed()){
+            return redirect(route('home'))->with('error', 'Deze mensa is gesloten en je kan je hiervoor dus niet meer uitschrijven.');
+        }
+
         // This is just a soft-delete. This is because we might need to retrieve information later.
         $mensaUser->intros()->delete();
         $mensaUser->delete();
